Fixed bug. Points must be passed in order.
When transferring points from the Point Map to the points buffer,
the points must be stored in the location indicated by the second
element of the map pairs. Those are the ordered locations of the
points, to which the cell point ids are referring.

Now manage to read and write the sphere.stl file.